fix(ci): python-coverage-shards — split correction fingerprint API over the 1500-line gate
Review fixes grew comment_verdict_residue to 1502 lines; move porcelain decode and correction fingerprint/mutation helpers into comment_verdict_residue_fingerprint and re-export them so the maintainability check passes again.
